+++ CHANGELOG.md
@@ -1,9 +1,14 @@
 # Changelog
 
-## 1.0.1 - WIP
+
+## 1.1.0 - 2025-10-19
+- Introducing the trait `FusionBaseModelTrait`
+
+## 1.0.1 - 2025-10-18
 - Adding Static Code Analysis via PHPStan + Larastan (Level 6)
 - Adding env variable FUSION_CONTENT_DIRECTORY for setting the path of content folder (default is app/resources/content)
 - Adding tests for hydrated model
+- Simplify and improving the DX for creating Models
 
 ## 1.0.0 - 2025-10-17
 - Upgrading the support for Laravel 12